# [PERFORMANCE] Improve vocab lookup performance by working with a dict() directly