@leezu 2020-10-05
This PR is currently blocked by an unrelated CI issue. I opened https://github.com/dmlc/gluon-nlp/pull/1383/ to attempt fixing it.